Revert "Merge branch 'oldschool' into 'dev'"
This reverts commit cf6894047e8e279dfcc90dfbae04f9d85f902b20

+++ OpenNos.DAL.EF/Entities/Map.cs
@@ -45,7 +45,7 @@ namespace OpenNos.DAL.EF
         public byte[] Data { get; set; }
 
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int MapId { get; set; }
+        public short MapId { get; set; }
 
         public virtual ICollection<MapMonster> MapMonster { get; set; }
