+++ OpenNos.DAL.EF/Entities/MapTypeMap.cs
@@ -24,7 +24,7 @@ namespace OpenNos.DAL.EF
         public virtual Map Map { get; set; }
 
         [Key, Column(Order = 0)]
-        public int MapId { get; set; }
+        public short MapId { get; set; }
 
         public virtual MapType MapType { get; set; }
