+++ OpenNos.Data/RespawnMapTypeDTO.cs
@@ -18,7 +18,7 @@ namespace OpenNos.Data
     {
         #region Properties
 
-        public int DefaultMapId { get; set; }
+        public short DefaultMapId { get; set; }
 
         public short DefaultX { get; set; }
