+++ OpenNos.GameObject/Map/Map.cs
@@ -32,7 +32,7 @@ namespace OpenNos.GameObject
 
         #region Instantiation
 
-        public Map(int mapId, byte[] data)
+        public Map(short mapId, byte[] data)
         {
             _random = new Random();
             MapId = mapId;
@@ -72,7 +72,7 @@ namespace OpenNos.GameObject
 
         public GridPos[,] Grid { get; set; }
 
-        public int MapId { get; set; }
+        public short MapId { get; set; }
 
         public List<MapTypeDTO> MapTypes { get; }
